opentalk/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
python-telegram-bot==20.7
Pillow==11.1.0
nest-asyncio==1.6.0 